requirements.txt
python-barcode[images]==0.16.1

